==> 1-data-understanding.ipynb <==
# IEEE-CIS Fraud Detection

## Data Understanding

# Import necessary modules
import pandas as pd
import numpy as np

### Importing Data

# Create folder path
folder_path = 'ieee-fraud-detection/'

# Read data
train_identity = pd.read_csv(f'{folder_path}train_identity.csv')
train_transaction = pd.read_csv(f'{folder_path}train_transaction.csv')
test_identity = pd.read_csv(f'{folder_path}test_identity.csv')
test_transaction = pd.read_csv(f'{folder_path}test_transaction.csv')
sub = pd.read_csv(f'{folder_path}sample_submission.csv')

def reduce_mem_usage(df, verbose=True):
    '''Returns DataFrame with reduced memory usage.'''
    numerics = ['int16', 'int32', 'int64', 'float16', 'float32', 'float64']
    start_mem = df.memory_usage().sum() / 1024**2    
    for col in df.columns:
        col_type = df[col].dtypes
        if col_type in numerics:
            c_min = df[col].min()
            c_max = df[col].max()
            if str(col_type)[:3] == 'int':
                if c_min > np.iinfo(np.int8).min and c_max < np.iinfo(np.int8).max:
                    df[col] = df[col].astype(np.int8)
                elif c_min > np.iinfo(np.int16).min and c_max < np.iinfo(np.int16).max:
                    df[col] = df[col].astype(np.int16)
                elif c_min > np.iinfo(np.int32).min and c_max < np.iinfo(np.int32).max:
                    df[col] = df[col].astype(np.int32)
                elif c_min > np.iinfo(np.int64).min and c_max < np.iinfo(np.int64).max:
                    df[col] = df[col].astype(np.int64)  
            else:
                if c_min > np.finfo(np.float16).min and c_max < np.finfo(np.float16).max:
                    df[col] = df[col].astype(np.float16)
                elif c_min > np.finfo(np.float32).min and c_max < np.finfo(np.float32).max:
                    df[col] = df[col].astype(np.float32)
                else:
                    df[col] = df[col].astype(np.float64)    
    return df

### Investigating Data

# Print the shape of the train data
train_identity.shape, train_transaction.shape

# Print the shape of the test data
test_identity.shape, test_transaction.shape

# Print column names
train_identity.columns.to_list()

train_identity.head()

train_transaction.head()

### Merging Data

# Merge datasets
train = pd.merge(train_transaction, train_identity, on='TransactionID', how='left')
test = pd.merge(test_transaction, test_identity, on='TransactionID', how='left')

train = reduce_mem_usage(train)
test = reduce_mem_usage(test)

train.shape, test.shape

# del train_identity, train_transaction, test_identity, test_transaction

# Columns with missing values
train.isnull().any().sum()

Almost all columns have missing data.

# Create a sample dataset for trials
train_sample = train.sample(1000)

train_sample.isnull().mean()[train_sample.isnull().mean() > .8]

# Print the columns with missing values higher than 80%
train.isnull().mean()[train.isnull().mean() > .8]

sub.head()

train.isFraud.value_counts()

# Print the mean of the isFraud column
train.isFraud.mean()

The data is highly imbalanced as expected.

### Column Groups

print(train.columns.to_list())

similar = ['Transaction', 'card', 'addr', 'dist', 'email', 
            'Product', '^C', '^(D[0-9])', '^M', '^V', 'id', 'Device']

similar_cols = [train.filter(regex=(i)).columns.to_list() for i in similar]


len(similar_cols)

print(similar_cols)

import pickle

# Save similar columns list as pickle
with open("dataset/similar_cols.txt", "wb") as cl:   #Pickling
    pickle.dump(similar_cols, cl)

# # Load similar columns list from memory
# with open("dataset/similar_cols", "rb") as cl:   # Unpickling
#     similar_cols = pickle.load(cl)

### Saving Data to `pickle`

pd.to_pickle(train, 'dataset/train.pkl')

pd.to_pickle(test, 'dataset/test.pkl')

pd.to_pickle(train, 'dataset/train_mem.pkl')

pd.to_pickle(test, 'dataset/test_mem.pkl')
