==> 3-missing-value-outlier-analysis.ipynb <==
# IEEE-CIS Fraud Detection

## Missing Value and Outlier Analysis

def corr_thresh(data, thres):
    '''
    Returns the corelation matrix with columns which have higher value than the threshold.
    '''
    # Create a correlation DataFrame
    corr = data.corr()
    
    # Change the self corelation from "1" to "0"
    for i in range(len(corr)):
        corr.iloc[i, i] = 0
    
    # Stack > select > unstack
    c = corr.stack()
    c_stacked = c[c.abs()>thres]
    c_unstacked = c_stacked.unstack(1)
    
    return c_unstacked

corr_new = corr_thresh(train[C_cols], 0.8)
corr_new.stack().sort_values()

def corr_drop_list(data, thres):
    '''
    Returns list of columns to drop
    '''
    # Create a correlation DataFrame
    corr = data.corr().abs()
    
    # Select upper triangle of correlation matrix
    corr_upper = corr.where(np.triu(np.ones(corr.shape), k=1).astype('bool'))
    
    # Find columns with correlation greater than the threshold
    drop_cols = [column for column in corr_upper.columns if any(corr_upper[column] > thres)]
    
    return drop_cols

drop_list = corr_drop_list(train[C_cols], thres = .8)

len(drop_list)

print(drop_list)
