==> 6-SMOTE.ipynb <==
# Import necessary modules

import pandas as pd
import numpy as np

%matplotlib inline
import matplotlib.pyplot as plt

import pickle

from sklearn.model_selection import train_test_split
from sklearn.decomposition import PCA
from sklearn.preprocessing import MinMaxScaler
from sklearn.preprocessing import StandardScaler

from sklearn.metrics import classification_report,  confusion_matrix
from sklearn.ensemble import RandomForestClassifier

# Import train data
df = pd.read_pickle('dataset/train_engineered_new.pkl')
test = pd.read_pickle('dataset/test_engineered_new.pkl')

## Evaluation Metrics

### Classification Metrics

from sklearn.metrics import classification_report, confusion_matrix
from sklearn.metrics import accuracy_score, roc_auc_score
from sklearn.metrics import average_precision_score, precision_recall_curve


def eval_metrics(actual, pred):
    print("Accuracy: ", accuracy_score(actual, pred))
    print (confusion_matrix(actual, pred))
    print (classification_report(actual, pred))
    
def model_results(X_train, y_train, X_test, y_test, model):
    # Fit the training model to the training set
    model.fit(X_train, y_train)
    
    # Obtain the predicted values and probabilities from the model 
    predicted = model.predict(X_test)
    probs = model.predict_proba(X_test)
    
    # Print the accuracy and roc_auc_score performance metric
    print("Accuracy: ", accuracy_score(y_test, predicted))
    print("AUC: ", roc_auc_score(y_test, probs[:,1]))
    print("-----")
    # Print the classification report and confusion matrix
    print (confusion_matrix(y_test, predicted))
    print (classification_report(y_test, predicted))

## Create X and y

### With Missing

# Create a list of columns to drop
remove_list = ['isFraud']

# Create features and target
X = df.drop(remove_list, axis=1)
X = pd.get_dummies(X)
y = df.isFraud

# Split data into training and test set
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=.3, random_state=0, stratify=y)

X.shape

### Without Missing

# Create a list of columns to drop
remove_list = ['isFraud']

# Create a list of columns with more than 20% missing
remove_missing_cols = df.isna().mean()[df.isna().mean()>0.2].index.to_list()
# df.drop(remove_missing_cols, axis=1, inplace=True)

no_missing = df.drop(remove_missing_cols, axis=1).dropna()

# Create features and target
X = no_missing.drop(remove_list, axis=1)
y = no_missing.isFraud

# Split data into training and test set
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=.3, random_state=0, stratify=y)

X.shape

## Pipeline

from sklearn.preprocessing import RobustScaler
from sklearn.decomposition import PCA

scaler = RobustScaler()
X_scaled = scaler.fit_transform(X)

pca = PCA(n_components=10)
p_components = pca.fit_transform(X_scaled)

print(pca.explained_variance_ratio_)
print(pca.explained_variance_ratio_.cumsum())

from sklearn.preprocessing import MinMaxScaler
from sklearn.decomposition import PCA
from imblearn.over_sampling import SMOTE
from sklearn.ensemble import RandomForestClassifier
# from sklearn.neighbors import KNeighborsClassifier as KNN
from imblearn.pipeline import Pipeline
#from sklearn.model_selection import GridSearchCV



# Initiate steps
smt = SMOTE(sampling_strategy="minority", random_state=42)
rf = RandomForestClassifier(n_estimators=200, max_depth=30, 
                            random_state=42)

# Create the pipeline
pipeline = Pipeline([('smt', smt), 
                     ('rf', rf)])

model_results(X_train, y_train, X_test, y_test, pipeline)

rf = RandomForestClassifier(n_estimators=400, max_depth=40, 
                            class_weight='balanced', random_state=42, 
                            n_jobs=-1)

model_results(X_train, y_train, X_test, y_test, rf)
